refactor: rename login folder
