+++ frontend/src/lib/components/login/LoginFooter.svelte
