+++ frontend/src/lib/components/login/LoginHeader.svelte
