+++ frontend/src/lib/components/login/LoginMenuItems.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { MenuItem, IconNorthEast, IconGitHub } from "@dfinity/gix-components";
   import { i18n } from "$lib/stores/i18n";
-  import LoginFooter from "$lib/components/auth/LoginFooter.svelte";
+  import LoginFooter from "$lib/components/login/LoginFooter.svelte";
 </script>
 
 <MenuItem
