+++ frontend/src/routes/(login)/+layout.svelte
@@ -3,9 +3,9 @@
   import { onMount } from "svelte";
   import { initAppAuth } from "$lib/services/$public/app.services";
   import { Layout, ContentBackdrop } from "@dfinity/gix-components";
-  import LoginMenuItems from "$lib/components/auth/LoginMenuItems.svelte";
-  import LoginFooter from "$lib/components/auth/LoginFooter.svelte";
-  import LoginHeader from "$lib/components/auth/LoginHeader.svelte";
+  import LoginMenuItems from "$lib/components/login/LoginMenuItems.svelte";
+  import LoginFooter from "$lib/components/login/LoginFooter.svelte";
+  import LoginHeader from "$lib/components/login/LoginHeader.svelte";
 
   onMount(async () => await initAppAuth());
 </script>
